Add login and auth wrapper with token support
