+++ lib/api.ts
@@ -15,7 +15,10 @@ export async function startReminders({
 }) {
   const res = await fetch(`${API_BASE}/start`, {
     method: "POST",
-    headers: { "Content-Type": "application/json" },
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${localStorage.getItem("token") ?? ""}`,
+    },
     body: JSON.stringify({
       phone_number,
       messages,
@@ -30,7 +33,10 @@ export async function startReminders({
 export async function stopReminders(phone_number: string) {
   const res = await fetch(`${API_BASE}/stop`, {
     method: "POST",
-    headers: { "Content-Type": "application/json" },
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${localStorage.getItem("token") ?? ""}`,
+    },
     body: JSON.stringify({ phone_number }),
   });
   return res.json();
@@ -39,13 +45,23 @@ export async function stopReminders(phone_number: string) {
 export async function markDone(phone_number: string) {
   const res = await fetch(`${API_BASE}/done`, {
     method: "POST",
-    headers: { "Content-Type": "application/json" },
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${localStorage.getItem("token") ?? ""}`,
+    },
     body: JSON.stringify({ phone_number }),
   });
   return res.json();
 }
 
 export async function getStatus(phone_number: string) {
-  const res = await fetch(`${API_BASE}/status?phone_number=${encodeURIComponent(phone_number)}`);
+  const res = await fetch(
+    `${API_BASE}/status?phone_number=${encodeURIComponent(phone_number)}`,
+    {
+      headers: {
+        Authorization: `Bearer ${localStorage.getItem("token") ?? ""}`,
+      },
+    }
+  );
   return res.json();
-} 
+}
